Rename package to meet go convention
